fstests: repository @ oss.sgi.com no longer used
We've not used the repository on oss.sgi.com as the primary
development target for some time. It's been keep up to date so that
old scripts didn't need to be updated.

We've just been warned that the future of oss.sgi.com is uncertain
as a result of the HPE buyout of SGI, and so with this we are
officially end-of-lifing the oss.sgi.com xfstests repository.

If you are reading this commiti message , it's probably because you
wondered why xfstests no longer built or ran. We broke the tree so
you would look to see what caused it and then read this commit....

The official location for xfstests source is now:

git://git.kernel.org/pub/scm/fs/xfs/xfstests-dev.git

Please point your scripts and test environments at this repository.

Signed-off-by: Dave Chinner <david@fromorbit.com>